Make it connecting even with old firmwares
When I got a new board Wemos Lolin D32 the example did not work as
the firmware is older than I had on my previous boards. To provide
a smoother start for people it was needed to get rid of network.STAT_*
constants, as they were missing in the older firmwares.